+++ CHANGELOG.md
@@ -10,9 +10,15 @@ All notable changes to this project will be documented in this file.
 - Openshift compatibility ([#323]).
 - Incorporated cluster-operation change. ([#323]).
 - Extend cluster resources for status and cluster operation (paused, stopped) ([#324]).
+- Cluster status conditions ([#326]).
 
 ### Changed
 
+- [BREAKING]: Support specifying Service type by moving `serviceType` (which was an experimental feature) to `clusterConfig.listenerClass`.
+  This enables us to later switch non-breaking to using `ListenerClasses` for the exposure of Services.
+  This change is breaking, because - for security reasons - we default to the `cluster-internal` `ListenerClass`.
+  If you need your cluster to be accessible from outside of Kubernetes you need to set `clusterConfig.listenerClass`
+  to `external-unstable` or `external-stable` ([#327]).
 - `operator-rs` `0.36.0` → `0.39.0` ([#324]).
 
 ### Fixes
@@ -23,6 +29,8 @@ All notable changes to this project will be documented in this file.
 [#317]: https://github.com/stackabletech/hive-operator/pull/317
 [#323]: https://github.com/stackabletech/hive-operator/pull/323
 [#324]: https://github.com/stackabletech/hive-operator/pull/324
+[#326]: https://github.com/stackabletech/hive-operator/pull/326
+[#327]: https://github.com/stackabletech/hive-operator/pull/327
 
 ## [23.1.0] - 2023-01-23